Se agrega lista de materiales solicitados a lista de control producción y se agrega buscar campos a admin de inventario

+++ Productos/models.py
@@ -252,7 +252,7 @@ class Productos_colores(models.Model):
         null=True,
         blank=True
     )
-    dosificacion = models.PositiveIntegerField(
+    dosificacion = models.FloatField(
         verbose_name='Dosificación',
         null=True,
         blank=True
